Better consistency between ApiKeys and UserApiKeys tables
