+++ core/src/org/labkey/core/query/ApiKeysUpdateService.java
@@ -8,22 +8,25 @@ import org.labkey.api.data.TableInfo;
 import org.labkey.api.dataiterator.DataIteratorBuilder;
 import org.labkey.api.query.BatchValidationException;
 import org.labkey.api.query.DefaultQueryUpdateService;
-import org.labkey.api.query.DuplicateKeyException;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.InvalidKeyException;
 import org.labkey.api.query.QueryUpdateServiceException;
-import org.labkey.api.query.ValidationException;
 import org.labkey.api.security.ApiKeyManager;
 import org.labkey.api.security.User;
+import org.labkey.api.security.permissions.UserManagementPermission;
 import org.labkey.api.view.UnauthorizedException;
 
 import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
 
-public class UserApiKeysUpdateService extends DefaultQueryUpdateService
+/**
+ * Delete is the only supported operation. Site and application admins can delete any API key; others are restricted to
+ * deleting their own.
+ */
+public class ApiKeysUpdateService extends DefaultQueryUpdateService
 {
-    public UserApiKeysUpdateService(TableInfo tableInfo, TableInfo dbTable)
+    public ApiKeysUpdateService(TableInfo tableInfo, TableInfo dbTable)
     {
         super(tableInfo, dbTable);
     }
@@ -31,13 +34,31 @@ public class UserApiKeysUpdateService extends DefaultQueryUpdateService
     @Override
     public List<Map<String, Object>> updateRows(User user, Container container, List<Map<String, Object>> rows, List<Map<String, Object>> oldKeys, BatchValidationException errors, @Nullable Map<Enum, Object> configParameters, Map<String, Object> extraScriptContext) throws InvalidKeyException, BatchValidationException, QueryUpdateServiceException, SQLException
     {
-        throw new UnsupportedOperationException("Updates are not allowed for table core.UserApiKeys.");
+        throw new UnsupportedOperationException("Updates are not allowed for this table.");
     }
 
     @Override
     protected Map<String, Object> updateRow(User user, Container container, Map<String, Object> row, @NotNull Map<String, Object> oldRow, boolean allowOwner, boolean retainCreation)
     {
-        throw new UnsupportedOperationException("Updates are not allowed for table core.UserApiKeys.");
+        throw new UnsupportedOperationException("Updates are not allowed for this table.");
+    }
+
+    @Override
+    public int importRows(User user, Container container, DataIteratorBuilder rows, BatchValidationException errors, @Nullable Map<Enum, Object> configParameters, Map<String, Object> extraScriptContext)
+    {
+        throw new UnsupportedOperationException("Import is not allowed for this table.");
+    }
+
+    @Override
+    protected Map<String, Object> insertRow(User user, Container container, Map<String, Object> row)
+    {
+        throw new UnsupportedOperationException("Insert is not allowed for this table.");
+    }
+
+    @Override
+    protected int truncateRows(User user, Container container)
+    {
+        throw new UnsupportedOperationException("Truncation of this table is not supported.");
     }
 
     @Override
@@ -46,7 +67,7 @@ public class UserApiKeysUpdateService extends DefaultQueryUpdateService
         if (oldRowMap == null)
             return null;
 
-        validateUser(user, oldRowMap);
+        validateDelete(user, oldRowMap);
 
         // We allow deletion of API keys from any container since they are defined globally,
         // so we skip the container permission check from the base class.
@@ -58,32 +79,16 @@ public class UserApiKeysUpdateService extends DefaultQueryUpdateService
         return oldRowMap;
     }
 
-    @Override
-    public int importRows(User user, Container container, DataIteratorBuilder rows, BatchValidationException errors, @Nullable Map<Enum, Object> configParameters, Map<String, Object> extraScriptContext)
+    protected void validateDelete(User user, Map<String, Object> row)
     {
-        throw new UnsupportedOperationException("Import is not allowed for table core.UserApiKeys.");
-    }
-
-    @Override
-    protected Map<String, Object> insertRow(User user, Container container, Map<String, Object> row) throws DuplicateKeyException, ValidationException, QueryUpdateServiceException, SQLException
-    {
-        // check if user is the same as the user in the row
-        validateUser(user, row);
-        return super.insertRow(user, container, row);
-    }
-
-    @Override
-    protected int truncateRows(User user, Container container)
-    {
-        throw new UnsupportedOperationException("Truncation of table core.UserApiKeys is not supported.");
-    }
-
-    protected void validateUser(User user, Map<String, Object> row)
-    {
-        Integer createdById = (Integer) row.get("CreatedBy");
-        if (createdById == null)
-            return;
-        if (user.getUserId() != createdById)
-            throw new UnauthorizedException("You do not have permission to delete the selected row.");
+        // Site and application admins can delete anyone's API key; others are restricted to deleting their own.
+        if (!user.hasRootPermission(UserManagementPermission.class))
+        {
+            Integer createdById = (Integer) row.get("CreatedBy");
+            if (createdById == null)
+                return;
+            if (user.getUserId() != createdById)
+                throw new UnauthorizedException("You do not have permission to delete the selected row.");
+        }
     }
 }
